Create a row component for match statistics
